// src/usecase/candidate/create/create-candidate.dto.ts
export type inputCandidateDTO = {
  id: string;
  name: string;
  email: string;
  phone: string;
  address: string;
  experience: string[];
  education: string[];
};

export type outputCandidateDTO = {
  id: string;
  name: string;
  email: string;
  phone: string;
  address: string;
  experience: string[];
  education: string[];
};